pre-login pages and login
Got pre-login page (index.php) with splash screen, signup, and login working.
Added "sansation" font family and used it for "vtrak" logo text.
Added error messages for invalid password or username for login and for non-matching passwords or already in use username for signup. Also repopuates the rest of the fields in the event of those errors so the user doesn't have to retype the other information.
Got login working using bcrypt hashing.
Fixed progress bar on dashboard.php to show 0 of goal upon first log in when hours completed is 0.